requirements-dev.txt
black==25.1.0
cfgv==3.4.0
click==8.1.8
colorama==0.4.6
distlib==0.3.9
filelock==3.18.0
flake8==7.2.0
identify==2.6.9
iniconfig==2.1.0
isort==6.0.1
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.7
pluggy==1.5.0
pre_commit==4.2.0
pycodestyle==2.13.0
pyflakes==3.3.2
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
pyreadline3==3.5.4
pyserial==3.5
pytest==8.3.5
PyYAML==6.0.2
virtualenv==20.30.0
